Mechanize promote and demote capability.
Now a stock can be promoted or demoted.

Usage:
To promote a stock, click the stock on the row and then
click "promote". Likewise to demote.  You cannot
promote a stock on the topmost row or demote a stock
on the lowest row.

Limitations:
Right now you must click the row you want to pro/de-mote
every single time.
Put in some stand-in test stocks in the WatchlistApp
constructor to avoid clicking during testing.

Future work:
Fix the limitation above.
Need to be able to save what is entered.