Add feature to save and load watchlist items to db
The save button will save the existing table out to an
sqllite database. Whatever is there just overwrites the
database.  When the app loads it will search if the
database file exists. If the file does exist, then
the database items are loaded and the watchlist
is regenerated.

Usage:
click save to save datbase to file
click load to load database from file

Limitations:
Note that the file name is hardcoded.

Future work:
Remove the load button and change its feature to be a
dialog such that the user can load any specified
database, Making this an issue to track