Add remaining stock info members, updates, cleanup
Based on pg 207 of Jason Kelly's book, Neatest Little
Guide to Stock Market Investing, have filled out the
Stock Info object with the watchlist items listed
on his worksheet. I have implemented it as two different
dataframes for ease of implementation to update the table
gui.

Usage:
no changes

Limitations:
Still fake data
Database does not really load the data back in. The info is
only regenerated from a couple fields on one of the now two
tables. Later will have to store the info if history of
metrics is to be accounted for.

Future work:
Need to make some of these error messages.
Identified need for another GUI widget. Something to show basic
info when a row is selected. Framework is set up so that can
be dropped in with some ease.